Fix: Remove unused dependencies and update MQTT connection logic
This commit removes unused dependencies from the project and updates the MQTT connection logic in `MqttService`.

The following dependencies were removed:

-   `com.squareup.retrofit2:retrofit`
-   `com.squareup.retrofit2:converter-gson`
-   `com.squareup.okhttp3:okhttp`
-   `com.squareup.okhttp3:logging-interceptor`
-   `com.github.bumptech.glide:glide`
-   `com.github.bumptech.glide:compiler`

These dependencies were not being used in the project and were adding unnecessary bloat.

The MQTT connection logic in `MqttService` was updated to remove the username and password authentication.
The device ID is now being used to generate a unique client ID and topic for each device.
This improves the security of the application by removing hardcoded credentials.

Additionally, the following changes were made:

-   Removed unused imports in `MainActivity`.
-   Removed unnecessary marquee effect for `messageTextView` in `MainActivity`.
-   Decreased the font size of `messageTextView` in `activity_main.xml`.
-   Updated the Android Gradle Plugin version to `8.7.3` in `libs.versions.toml`.